src/main/java/song/mygg/domain/common/service/RestService.java
package song.mygg.domain.common.service;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.*;
import org.springframework.stereotype.Service;
import org.springframework.web.client.HttpClientErrorException;
import org.springframework.web.client.RestTemplate;
import song.mygg.domain.common.entity.CommonEntity;

import java.util.Optional;

@Slf4j
@Service
@RequiredArgsConstructor
public class RestService {
    private final RestTemplate restTemplate;

    public <T extends CommonEntity> Optional<T> findEntity(String url, Class<T> clazz) throws HttpClientErrorException {
        HttpHeaders headers = new HttpHeaders();
        headers.setContentType(MediaType.APPLICATION_JSON);
        HttpEntity<T> httpEntity = new HttpEntity<>(headers);

        try {
            ResponseEntity<T> response = restTemplate.exchange(
                    url,
                    HttpMethod.GET,
                    httpEntity,
                    clazz
            );

            return Optional.ofNullable(response.getBody());
        } catch (HttpClientErrorException.NotFound e) {
            return Optional.empty();
        } catch (Exception e) {
            return Optional.empty();
        }
    }
}
